Simplify server implementation
