+++ packages/styletron-engine-atomic/src/server/server.js
@@ -2,8 +2,6 @@
 
 import SequentialIDGenerator from "../sequential-id-generator.js";
 
-import {generateHtmlString} from "./utils.js";
-
 import type {StandardEngine} from "styletron-standard";
 
 import {Cache, MultiCache} from "../cache.js";
@@ -118,6 +116,35 @@ class StyletronServer implements StandardEngine {
       this.fontFaceRules + stringify(this.styleRules) + this.keyframesRules
     );
   }
+}
+
+export function generateHtmlString(sheets: Array<sheetT>, className: string) {
+  let html = "";
+  for (let i = 0; i < sheets.length; i++) {
+    const sheet = sheets[i];
+    const {class: originalClassName, ...rest} = sheet.attrs;
+    const attrs = {
+      class: originalClassName
+        ? `${className} ${originalClassName}`
+        : className,
+      ...rest
+    };
+    html += `<style${attrsToString(attrs)}>${sheet.css}</style>`;
+  }
+  return html;
+}
+
+function attrsToString(attrs) {
+  let result = "";
+  for (const attr in attrs) {
+    const value = attrs[attr];
+    if (value === true) {
+      result += " " + attr;
+    } else if (value !== false) {
+      result += ` ${attr}="${value}"`;
+    }
+  }
+  return result;
 }
 
 function stringify(styleRules) {
