--- packages/styletron-engine-atomic/src/server/utils.js
@@ -1,32 +1,0 @@
-// @flow strict
-
-import type {sheetT} from "./server.js";
-
-export function generateHtmlString(sheets: Array<sheetT>, className: string) {
-  let html = "";
-  for (let i = 0; i < sheets.length; i++) {
-    const sheet = sheets[i];
-    const {class: originalClassName, ...rest} = sheet.attrs;
-    const attrs = {
-      class: originalClassName
-        ? `${className} ${originalClassName}`
-        : className,
-      ...rest
-    };
-    html += `<style${attrsToString(attrs)}>${sheet.css}</style>`;
-  }
-  return html;
-}
-
-function attrsToString(attrs) {
-  let result = "";
-  for (const attr in attrs) {
-    const value = attrs[attr];
-    if (value === true) {
-      result += " " + attr;
-    } else if (value !== false) {
-      result += ` ${attr}="${value}"`;
-    }
-  }
-  return result;
-}
